Enabled channel broadcasts for mod info
By default, commands for mod info will be sent via query ('.modname').
However, if the word 'show' is also used ('.modname show'), the info
will be broadcast to the whole channel. This will be useful for pointing
someone to info for a mod. Also available is '.mods' to display the full
list of modinfo commands to avoid flooding with '.help'