New Feature: Secret Links
Secret Links are a separate subset of links in the config that will act
as commands, but behave in a different manner. Secret commands will not
show up in '.commands' and they will only reply via query to the person
issuing the command. This is just a fun exercise of what snaibot is
capable of. This may be expanded in the future with additional features
and commands can be added at any time to the config, just like regular
links.